PR corrections: Changed the logic of the save state and persistence
