--- react/game/src/utils/application.js
@@ -1,18 +1,0 @@
-export const loadState = () => {
-  try {
-    const serializedState = localStorage.getItem('gameState');
-    if (serializedState === null) return undefined;
-    return JSON.parse(serializedState);
-  } catch (err) {
-    return undefined;
-  }
-};
-
-export const saveState = gameState => {
-  try {
-    const serializedState = JSON.stringify(gameState);
-    localStorage.setItem('gameState', serializedState);
-  } catch (err) {
-    // Ignore write errors
-  }
-};
